embedding-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
fastembed==0.7.4
pydantic==2.10.4
python-dotenv==1.0.1
numpy==2.2.1
# Phase E: AI Agents
langgraph==1.0.7
langchain>=0.3.0
langchain-anthropic>=0.3.0
crewai==1.8.1
anthropic>=0.40.0
supabase>=2.0.0
httpx>=0.27.0
